Fix integer overflow in WAIT and WAITAOF numreplicas parameter
The numreplicas parameter in WAIT and WAITAOF commands is parsed as
`long` but stored in `blockingState.numreplicas` which is `int`. When a
value larger than INT_MAX (e.g., 2147483648) is passed, the implicit
narrowing conversion from `long` to `int` causes overflow, resulting in
a negative value that breaks the blocking logic.

Example: If `wait 2147483648 0` the client enters a blocked state, it
will unblock immediately after receiving ACKs from the replicas since
`numreplicas` is a negative value, we consider the `numreplicas`
requirement to be satisfied.

It's not a major issue, but it is indeed a bug.

Additionally, the WAIT command used `getLongFromObjectOrReply()` which
accepted negative values for numreplicas. A negative number should be
an invalid value. And in WAITAOF command we will reject negative number.

It is not a major issue, it is an inconsistency in error handling between
WAIT and WAITAOF.

Fix both issues by using `getRangeLongFromObjectOrReply()` with range
[0, INT_MAX] for numreplicas validation in both WAIT and WAITAOF. This
ensures:
- Negative values are rejected with an error.
- Values exceeding INT_MAX are rejected, preventing integer overflow.
- Both commands now have consistent validation for numreplicas.

This could be a potentially breaking change. In the past, we allowed:
- WAIT allows negative numreplicas values.
- WAIT and WAITAOF allows numreplicas to exceed INT_MAX.

Add tests for numreplicas boundary validation in both commands, and
add missing parameter validation tests for WAITAOF (numlocal range,
numreplicas range, and timeout range).

+++ src/server.h
@@ -3880,7 +3880,7 @@ void signalKeyAsReady(serverDb *db, robj *key, int type);
 void blockForKeys(client *c, int btype, robj **keys, int numkeys, mstime_t timeout, int unblock_on_nokey);
 void blockClientShutdown(client *c);
 void blockPostponeClient(client *c);
-void blockClientForReplicaAck(client *c, mstime_t timeout, long long offset, long numreplicas, int numlocal);
+void blockClientForReplicaAck(client *c, mstime_t timeout, long long offset, int numreplicas, int numlocal);
 void replicationRequestAckFromReplicas(void);
 void signalDeletedKeyAsReady(serverDb *db, robj *key, int type);
 void updateStatsOnUnblock(client *c, long blocked_us, long reply_us, int failed_or_rejected);
